moved generic getData to new sagas folder
